last comment in 2023 from Doston

+++ app/Http/Controllers/CropsGenerationController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Area;
+use App\Models\CropData;
 use App\Models\Region;
 use App\Models\CropsGeneration;
 use App\tbl_states;
@@ -53,6 +54,10 @@ class CropsGenerationController extends Controller
     public function destory($id)
     {
         $this->authorize('setting_delete', User::class);
+        $app = CropData::where('generation_id',$id)->first();
+        if($app){
+            return redirect('crops_generation/list')->with('message', 'Cannot Deleted');
+        }
         CropsGeneration::destroy($id);
         return redirect('crops_generation/list')->with('message', 'Successfully Deleted');
     }
